frontend/requirements.txt
Django==4.2.10
requests==2.31.0
python-dotenv==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
gunicorn==21.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
psycopg2-binary==2.9.9 